Modifications for the following mxml syntax: <schema:WorldMapElement id="myMap" styleName="grad1" targetLongProjection = "{latProj}"> 	<schema:targetLatProjection> 		<trans:Projection id="latProj" proj="Miller cylindrical" /> 	</schema:targetLatProjection> </schema:WorldMapElement>
Appended an implementation of commitProperties to the end of class WorldMapElement, which adds all country polygons to WorldMapElement. Also created a class birdeye.vis.trans.projections.Projection that handles all the transformations and modified the transformations so that separate functions are called for calculating x and y values. The calculation takes parameters size, minLat, maxLat, minLong and maxLong in order to allow the caller to decide size and zoom for the calculated map.